refactor: export data-related stores from a data-store file
