+++ src/store/data-store.js
@@ -1,0 +1,4 @@
+import { Dataframe } from './dataframe';
+
+export const dataTable = new Dataframe();
+export const infoTable = new Dataframe();
